Move tests. Use local XML.
