+++ Cargo.toml
@@ -21,7 +21,7 @@ serde = { version = "1.0.164", features = ["derive"] }
 serde_repr = "0.1.12"
 quick-xml = { version = "0.29.0", features = ["serialize"]}
 atoi = "2.0.0"
+reqwest = { version = "0.11.18", features = ["blocking"] }
 
 # standard crate data is left out
 [dev-dependencies]
-reqwest = { version = "0.11.18", features = ["blocking"] }
