+++ src/helpers.rs
@@ -1,3 +1,4 @@
+use std::fs;
 use quick_xml::events::Event;
 use quick_xml::Reader;
 use crate::seq::BioSeq;
@@ -99,9 +100,11 @@ pub fn build_fetch_url(db: EntrezDb, id: &str, r#type: &str, mode: &str) -> Stri
 
 pub enum DataType {
     BioSeqSet(BioSeqSet),
+    /// placeholder for other types
+    EtAl
 }
 
-pub fn parse_response(response: &str) -> Result<DataType, ()> {
+pub fn parse_xml(response: &str) -> Result<DataType, ()> {
     let mut reader = Reader::from_str(response);
     loop {
         match reader.read_event().unwrap() {
@@ -120,4 +123,65 @@ pub fn parse_response(response: &str) -> Result<DataType, ()> {
         }
     }
     return Err(())
+}
+
+pub fn get_local_xml(path: &str) -> String {
+    let file = fs::read(path);
+    return
+        file.unwrap()
+            .escape_ascii()
+            .to_string()
+}
+
+pub fn fetch_data(db: EntrezDb, id: &str, r#type: &str, mode: &str) -> DataType {
+    let url = build_fetch_url(db, id, r#type, mode);
+    let response =
+        reqwest::blocking::get(url)
+            .unwrap()
+            .text()
+            .unwrap();
+    parse_xml(response.as_str()).unwrap()
+}
+
+#[cfg(test)]
+mod tests {
+    use std::fs;
+    use crate::{build_fetch_url, build_search_url, DataType, EntrezDb, get_local_xml, parse_xml};
+
+    #[test]
+    fn search_url() {
+        let url = build_search_url(EntrezDb::Protein, "deaminase");
+    }
+
+    #[test]
+    fn test_protein() {
+        let id = "2520667272";
+        let url = build_fetch_url(EntrezDb::Protein, id, "native", "xml");
+    }
+
+    #[test]
+    fn test_parse_xml() {
+        let data = get_local_xml("tests/data/2519734237.xml");
+        match parse_xml(data.as_str()).unwrap() {
+            DataType::BioSeqSet(_) => (),
+            _ => assert!(false)
+        }
+    }
+
+
+    #[test]
+    fn test_article_set() {
+        let id = "37332098";
+        let db = EntrezDb::PubMed;
+
+        let url = build_fetch_url(db, id, "xml", "xml");
+        let _ =
+            reqwest::blocking::get(url)
+                .unwrap()
+                .text()
+                .unwrap();
+        //let expected = from_str(text.as_str()).unwrap();
+        //assert!(expected.is_empty().not())
+    }
+
 }
